VIDSOL-346 fix video flickering on resize
